[fix] UseCaseを使用するように変更
